// src/app/containers/general-profile/general-profile.interface.ts
export interface GeneralProfile{
    name: string;
    image?: File;
    gender: string;
    nationality: string;
    cnic:string;
    dob: string;
    fname: string;
    fstatus: string;
    fprofession: string;
    fincome: string;
    mname: string;
    mstatus: string;
    mprofession: string;
    mincome: string
    religion: string;
    area: string;
    mnumber: string;
    address: string;
    board: string;
    subject: string;
}